requirements.txt
llama-index-core==0.11.0
llama_index-vector_stores-faiss==0.2.1
llama-index-llms-openai==0.2.0
llama-index-llms-azure-openai==0.2.0
llama-index-readers-web==0.2.0
llama-index-embeddings-huggingface==0.3.1
openai==1.63.2
faiss-cpu==1.10.0
pymupdf4llm==0.0.17
sentence_transformers==3.4.1
httpx==0.28.1
mammoth==1.9.0
markdownify==0.14.1
pyyaml==6.0.1
tqdm==4.66.5
tabulate==0.9.0
beautifulsoup4==4.13.3
playwright==1.46.0